+++ src/interfaces/index.ts
@@ -1,2 +1,3 @@
 export * from './dateTime/datetime.interface';
 export * from './dateTimeInput/dateTimeInput.interface';
+export * from './formatProvider/formatProvider.interface';