+++ src/legacy/picker/components/pickerBase.component.ts
@@ -2,7 +2,7 @@ import {ChangeDetectorRef, Directive, Inject} from '@angular/core';
 
 import {DateTimeValue} from '../../../interfaces/dateTime/datetime.interface';
 import {DATE_API} from '../../../misc/tokens';
-import {DateApi, DateApiObject} from '../../../services/dateApi.interface';
+import {DateApi, DateApiObject} from '../../../services/dateApi/dateApi.interface';
 import {DateTimePicker, PeriodData} from '../misc/datetimePicker.interface';
 import {PickerImplBaseComponent} from './pickerImplBase.component';