+++ src/legacy/picker/components/yearPicker/yearPicker.component.ts
@@ -2,7 +2,7 @@ import {Component, ChangeDetectionStrategy, Inject, ChangeDetectorRef} from '@an
 import {extend} from '@jscrpt/common';
 
 import {DATE_API} from '../../../../misc/tokens';
-import {DateApi, DateApiObject} from '../../../../services/dateApi.interface';
+import {DateApi, DateApiObject} from '../../../../services/dateApi/dateApi.interface';
 import {DateTimePicker, PeriodData, YearData} from '../../misc/datetimePicker.interface';
 import {PickerBaseComponent} from '../pickerBase.component';
 import {YearPickerCssClasses} from './yearPicker.interfaces';