+++ src/legacy/selector/components/selector/selector.component.ts
@@ -8,12 +8,13 @@ import {DATE_API, FORMAT_PROVIDER} from '../../../../misc/tokens';
 import {DateTimeDayPickerComponent} from '../../../picker/components/dayPicker/dayPicker.component';
 import {DateTimeMonthPickerComponent} from '../../../picker/components/monthPicker/monthPicker.component';
 import {DateTimeYearPickerComponent} from '../../../picker/components/yearPicker/yearPicker.component';
-import {DateApi, FormatProvider} from '../../../../services';
+import {DateApi} from '../../../../services';
 import {DateValueProvider} from '../../../../services/dateValueProvider.service';
 import {DateTimeSelector, DateTimeSelectorOptions} from '../../misc/datetimeSelector.interface';
 import {DATE_TIME_SELECTOR_CONFIGURATION} from '../../misc/tokens';
 import {InputDateTimeSelectorComponent} from '../inputDateTime/inputDateTime.component';
 import {enterLeaveAnimateChildTrigger} from './selector.component.animations';
+import {FormatProvider} from '../../../../interfaces';
 
 //TODO - add support for body absolute picker