+++ src/misc/enums.ts
@@ -1,0 +1,25 @@
+/**
+ * Available formats for date time value
+ */
+export enum DateTimeValueFormat
+{
+    /**
+     * Instance of date
+     */
+    DateInstance,
+
+    /**
+     * Numeric unix timestamp in seconds
+     */
+    UnixTimestamp,
+
+    /**
+     * Formatted date as string value
+     */
+    FormattedString,
+
+    /**
+     * Range of date instances from, to
+     */
+    RangeOfDateInstances,
+}