+++ src/pipes/dateFormat.pipe.ts
@@ -1,8 +1,9 @@
 import {Inject, Pipe, PipeTransform} from '@angular/core';
 import {nameof} from '@jscrpt/common';
 
+import {FormatProvider} from '../interfaces';
 import {DATE_API, FORMAT_PROVIDER} from '../misc/tokens';
-import {DateApi, DateValue, FormatProvider} from '../services';
+import {DateApi, DateValue} from '../services';
 
 /**
  * Pipe that is used for formatting date