+++ src/services/dateTimeRelativeParser.service.ts
@@ -2,7 +2,7 @@ import {Injectable, Injector} from '@angular/core';
 import {isString} from '@jscrpt/common';
 
 import {DATE_API} from '../misc/tokens';
-import {DateApi, DateApiObject, DateValue} from './dateApi.interface';
+import {DateApi, DateApiObject, DateValue} from './dateApi/dateApi.interface';
 
 /**
  * Service used for parsing relative date time strings