+++ src/services/index.ts
@@ -1,3 +1,2 @@
-export * from './dateApi.interface';
-export * from './formatProvider.interface';
+export * from './dateApi/dateApi.interface';
 export * from './datePositionParser.interface';